bugs and fixes

+++ queueUsingArray.cpp
@@ -3,39 +3,49 @@ using namespace std;
 
 const int MAX_SIZE = 100;
 
-class Queue {
+class Queue
+{
 private:
     int front, rear, size;
     int arr[MAX_SIZE];
 
 public:
-    Queue() {
+    Queue()
+    {
         front = -1;
         rear = -1;
         size = 0;
     }
 
-    bool isEmpty() {
+    bool isEmpty()
+    {
         return front == -1 && rear == -1;
     }
 
-    bool isFull() {
+    bool isFull()
+    {
         return (rear + 1) % MAX_SIZE == front;
     }
 
-    int getSize() {
+    int getSize()
+    {
         return size;
     }
 
-    void enqueue(int data) {
-        if (isFull()) {
+    void enqueue(int data)
+    {
+        if (isFull())
+        {
             cout << "Queue is full" << endl;
             return;
         }
 
-        if (isEmpty()) {
+        if (isEmpty())
+        {
             front = rear = 0;
-        } else {
+        }
+        else
+        {
             rear = (rear + 1) % MAX_SIZE;
         }
 
@@ -43,18 +53,23 @@ public:
         size++;
     }
 
-    int dequeue() {
-        if (isEmpty()) {
+    int dequeue()
+    {
+        if (isEmpty())
+        {
             cout << "Queue is empty" << endl;
-            return INT_MIN;
+            return INT16_MIN;
         }
 
         int data = arr[front];
         size--;
 
-        if (front == rear) {
+        if (front == rear)
+        {
             front = rear = -1;
-        } else {
+        }
+        else
+        {
             front = (front + 1) % MAX_SIZE;
         }
